% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_to_test.R
\name{calculate_minors}
\alias{calculate_minors}
\title{Calculate all minors of a specific size of a matrix}
\usage{
calculate_minors(A, size)
}
\arguments{
\item{A}{A matrix, whose minors to compute. The format of the entries of the matrix needs to be "m1_2_2"}

\item{size}{size of the minors}
}
\value{
List \code{all_minors} where each entry corresponds to one polynomial
}
\description{
Calculate all minors of a specific size of a matrix
}
