% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_to_test.R
\name{construct_flattening_matrix}
\alias{construct_flattening_matrix}
\title{Constructs (k/2)th flattening matrix of a tensor in (\mathbb{R}^p)^{\otimes k}.}
\usage{
construct_flattening_matrix(p, k)
}
\arguments{
\item{p}{Integer, number of nodes.}

\item{k}{Integer, order of tensor.}
}
\value{
Matrix fl.
}
\description{
Constructs (k/2)th flattening matrix of a tensor in (\mathbb{R}^p)^{\otimes k}.
}
