% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_to_test.R
\name{construct_M}
\alias{construct_M}
\title{Constructs matrix M}
\usage{
construct_M(p, orders)
}
\arguments{
\item{p}{Integer, number of nodes.}

\item{orders}{List of integers, orders of cumulants to include.}
}
\value{
Matrix M containing the cumulants of the lowest order in the first and the reamaining cumulants in 
the remaining rows.
}
\description{
Constructs matrix M
}
