% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_to_test.R
\name{construct_Y_T}
\alias{construct_Y_T}
\title{Constructs Young flattening Y_T}
\usage{
construct_Y_T(p, k)
}
\arguments{
\item{p}{Integer, number of nodes.}

\item{k}{Integer, order of the tensor.}
}
\value{
Young flattening Y_T as matrix.
}
\description{
Constructs Young flattening Y_T
}
