% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRT_statistic.R
\name{CRT}
\alias{CRT}
\title{CRT statistic}
\usage{
CRT(X, Pi, r, J = NULL, moment_expressions = NULL)
}
\arguments{
\item{X}{Matrix of size n x p containing the observed data.}

\item{Pi}{Matrix for that we test whether rank(Pi) <= 4r.}

\item{r}{Integer, rank bound.}

\item{J}{Matrix, Jacobian of Pi needed for calculating the asymtotic covariance of Pi using the delta method. Optional.}

\item{moment_expressions}{List of moment expressions appearing in Pi. Optional.}
}
\value{
List containing the p-value, the test statistic, the estimated Pi and its estimated covariance matrix.
}
\description{
CRT statistic
}
