% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generates data under the null hypothetis and different alternatives}
\usage{
generate_data(
  n,
  p,
  l,
  noise_distr,
  scenario = c("H_0", "alt_cos", "alt_quadr", "alt_conf"),
  delta = NULL
)
}
\arguments{
\item{n}{Integer, sample size.}

\item{p}{Integer, number of nodes.}

\item{l}{Integer, number of confounders}

\item{noise_distr}{String, distribution of noise terms.}

\item{scenario}{String, scenario to sample from.}

\item{delta}{Real Number, delta for local alternatives.}
}
\value{
Matrix with \code{n} rows and \code{p} columns. Each row corresponds 
to an independent sample of the distribution of X.
}
\description{
Generates data under the null hypothetis and different alternatives
}
