% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incomplete_U_statistic.R
\name{incomplete_U_stat}
\alias{incomplete_U_stat}
\title{Incomplete U-statistic}
\usage{
incomplete_U_stat(
  X,
  equality_constraints = list(),
  ineq_constraints = list(),
  E = 1000,
  n1 = min(nrow(X), 500),
  N = 2 * nrow(X)
)
}
\arguments{
\item{X}{Matrix of size n x p containing the observed data.}

\item{equality_constraints}{List containing all equalities to assess.}

\item{ineq_constraints}{List containing all inequalities to assess.}

\item{E}{Integer, number of bootstrap replications.}

\item{n1}{Integer, number of samples to take into account for estimating G}

\item{N}{Integer, computational budget parameter as described in the supplementary material.}
}
\value{
List containing the p-value and the test statistic.
}
\description{
Incomplete U-statistic
}
