\name{TestLinearSEM-package}
\alias{TestLinearSEM-package}
\alias{TestLinearSEM}
\docType{package}
\title{
  Goodness-of-fit tests for linear non-Gaussian structural equation model.
}
\description{
  This package provides the three different options to test the goodness-of-fit of a linear structural equation model as described in the paper Goodness-of-fit tests for linear non-Gaussian structural equation model. 
}
\details{
  Examples on how to use the package can be found in the vignette.
}
\author{
Daniela Schkoda

Maintainer: Daniela Schkoda <daniela.schkoda@tum.de>
}